Task2 for JS Training Center
A simple project, which lets you input song name and author and add it to the list. You can also clear the list, feature of removing single song is yet to be developed)

This project uses imports, Promises, async/await, classes.

in index1.js file there is an explanation of closure(scope?) of function canProceed(). The explanation states global lexical environment and both inner and outer lexical environments, i hope it was the point of this task)

when adding a song to the list, a new SongItem object is created and it is added to the SongList object